Refactor attribute renaming out from shape inference
