+++ onnx/common/visitor.h
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#pragma once
 #include "onnx/common/common.h"
 #include "onnx/onnx_pb.h"
